GROSSE UPDATE ; mode 64 bits (limine pour l'instant), récriture comprete des interruption, gdt, tss et autres en 64 bits, début d'un moteur de rendu 2d, ssaa, sqrt en libc et framebuffer

+++ kernel/idt/gdt.h
@@ -14,10 +14,10 @@ typedef struct __attribute__((packed)) {
 
 typedef struct __attribute__((packed)) {
     uint16_t limit;  
-    uint32_t base;   
+    uint64_t base;   
 } gdt_descriptor_t;
 
 void gdt_init(void);
-void gdt_set_tss_entry(uint32_t base, uint32_t limit);
+void gdt_set_tss_entry(uint64_t base, uint64_t limit);
 
 #endif
